Split text-color into UMD and ESM versions
