+++ addons/editor-theme3/theme3.js
@@ -1,4 +1,4 @@
-import { textColor, multiply } from "../../libraries/common/cs/text-color.js";
+import { textColor, multiply } from "../../libraries/common/cs/text-color.esm.js";
 
 function updateSettings(addon, newStyle) {
   var stylesheet = "";
