+++ libraries/common/cs/text-color.esm.js
@@ -1,0 +1,3 @@
+import * as _textColor from "./text-color.js";
+const { textColor, multiply, brighten } = globalThis.__scratchAddonsTextColor;
+export { textColor, multiply, brighten };
