+++ manifest.json
@@ -22,7 +22,7 @@
     {
       "matches": ["https://scratch.mit.edu/*"],
       "run_at": "document_start",
-      "js": ["libraries/thirdparty/cs/comlink.js", "content-scripts/cs.js"],
+      "js": ["libraries/thirdparty/cs/comlink.js", "libraries/common/cs/text-color.js", "content-scripts/cs.js"],
       "all_frames": true
     },
     {
